+++ src/__generated__/Fair_fair.graphql.ts
@@ -1,0 +1,34 @@
+/* tslint:disable */
+
+import { ConcreteFragment } from "relay-runtime";
+export type Fair_fair = {
+    readonly name: string | null;
+};
+
+
+
+const node: ConcreteFragment = {
+  "kind": "Fragment",
+  "name": "Fair_fair",
+  "type": "Fair",
+  "metadata": null,
+  "argumentDefinitions": [],
+  "selections": [
+    {
+      "kind": "ScalarField",
+      "alias": null,
+      "name": "name",
+      "args": null,
+      "storageKey": null
+    },
+    {
+      "kind": "ScalarField",
+      "alias": null,
+      "name": "__id",
+      "args": null,
+      "storageKey": null
+    }
+  ]
+};
+(node as any).hash = '5496c758924be928482513d395e6d47d';
+export default node;