+++ src/lib/AppRegistry.tsx
@@ -2,12 +2,13 @@ import React from "react"
 import { AppRegistry, View } from "react-native"
 
 import Consignments from "./Components/Consignments"
-import Containers from "./Containers/index"
+import * as Containers from "./Containers/index"
 import {
   ArtistRenderer,
   BidderFlowRendererProps,
   BidFlowRenderer,
   ConversationRenderer,
+  FairRenderer,
   GeneRenderer,
   InboxRenderer,
   InquiryRenderer,
@@ -122,6 +123,20 @@ const BidderFlow: React.SFC<BidderFlowProps> = ({ intent, ...restProps }) => {
   return <Renderer {...restProps} render={renderWithLoadProgress(Container)} />
 }
 
+interface FairProps {
+  fairID: string
+}
+
+const Fair: React.SFC<FairProps> = track<FairProps>(props => {
+  return {
+    context_screen: Schema.PageNames.FairPage,
+    context_screen_owner_slug: props.fairID,
+    context_screen_owner_type: Schema.OwnerEntityTypes.Fair,
+  }
+})(({ fairID }) => {
+  return <FairRenderer fairID={fairID} render={renderWithLoadProgress(Containers.Fair, { fairID })} />
+})
+
 AppRegistry.registerComponent("Consignments", () => Consignments)
 AppRegistry.registerComponent("Artist", () => Artist)
 AppRegistry.registerComponent("Home", () => HomeScene)
@@ -136,3 +151,4 @@ AppRegistry.registerComponent("Inquiry", () => Inquiry)
 AppRegistry.registerComponent("Favorites", () => FavoritesScene)
 // TODO: Change everything to BidderFlow? AuctionAction?
 AppRegistry.registerComponent("BidFlow", () => BidderFlow)
+AppRegistry.registerComponent("Fair", () => Fair)