+++ CHANGELOG.md
@@ -16,6 +16,8 @@
 
 ### Master
 
+- Adds a new ARComponentVC and root React Component for Fairs - orta
+
 ### 1.6.1
 
 - Tweaks to artwork badges - matt